chore: add libs checker utils
