+++ apps/server/src/routes/v1/monitors/trigger/post.ts
@@ -1,4 +1,5 @@
 import { env } from "@/env";
+import { getCheckerPayload, getCheckerUrl } from "@/libs/checker";
 import { OpenStatusApiError, openApiErrorResponses } from "@/libs/errors";
 import { createRoute, z } from "@hono/zod-openapi";
 import { and, eq, gte, isNull, sql } from "@openstatus/db";
@@ -10,7 +11,6 @@ import { monitor } from "@openstatus/db/src/schema/monitors/monitor";
 import { selectMonitorSchema } from "@openstatus/db/src/schema/monitors/validation";
 import { HTTPException } from "hono/http-exception";
 import type { monitorsApi } from "..";
-import { getCheckerPayload, getCheckerUrl } from "../utils";
 import { ParamsSchema, TriggerSchema } from "./schema";
 
 const postRoute = createRoute({
